backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pgvector==0.2.4
pydantic==2.5.3
pydantic-settings==2.1.0
httpx==0.26.0
python-dotenv==1.0.0
google-generativeai==0.4.0
alembic==1.13.1
asyncpg==0.29.0
beautifulsoup4==4.12.3
tenacity==8.2.3
numpy==1.26.3
sentence-transformers==2.2.2
